colab_requirements.txt
pytorch-lightning
git+https://github.com/albumentations-team/albumentations
bounding_box